refactor: only expose get active session interface
