+++ features/session/internal-api/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/session/internalapi/GetActiveSessionUseCase.kt
@@ -1,0 +1,5 @@
+package uk.gov.onelogin.criorchestrator.features.session.internalapi
+
+interface GetActiveSessionUseCase {
+    suspend fun execute(): Boolean
+}
