+++ features/session/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/session/internal/SessionApi.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.features.session.internalapi
+package uk.gov.onelogin.criorchestrator.features.session.internal
 
 import uk.gov.android.network.api.ApiResponse
