+++ features/session/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/session/internal/SessionStore.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.features.session.internalapi
+package uk.gov.onelogin.criorchestrator.features.session.internal
 
 import kotlinx.coroutines.flow.StateFlow
